+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 - Create Transactions associated with a Fund
 - Remove the distinction between Fund and Activity from the user
 - Users can edit an organisation
+- Users can edit the basic fund record